refactor: ProductService, StockManager 분리 - 순환참조 방지
