+++ src/main/java/com/flab/just_10_minutes/Product/service/ProductService.java
@@ -7,7 +7,6 @@ import com.flab.just_10_minutes.User.domain.User;
 import com.flab.just_10_minutes.User.infrastructure.UserDao;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -21,15 +20,5 @@ public class ProductService {
 
         Product product = ProductDto.toDomain(productDto, seller);
         return productDao.save(product);
-    }
-
-    @Transactional
-    public Product decreaseStock(final Long productId, final Long requestQuantity) {
-        Product product = productDao.fetchWithLock(productId);
-        Long updatedStock = Math.abs(requestQuantity) + product.getPurchasedStock();
-        if (updatedStock > product.getTotalStock()) {
-            throw new RuntimeException("over totalStock");
-        }
-        return productDao.patchStock(productId, updatedStock);
     }
 }
